Fix Checkbox Group disabled status (#7266)
