+++ components/checkbox/__tests__/checkbox.test.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import Checkbox from '..';
+
+describe('Checkbox', () => {
+  it('responses hover events', () => {
+    const onMouseEnter = jest.fn();
+    const onMouseLeave = jest.fn();
+
+    const wrapper = shallow(
+      <Checkbox
+        onMouseEnter={onMouseEnter}
+        onMouseLeave={onMouseLeave}
+      />
+    );
+
+    wrapper.simulate('mouseenter');
+    expect(onMouseEnter).toHaveBeenCalled();
+
+    wrapper.simulate('mouseleave');
+    expect(onMouseLeave).toHaveBeenCalled();
+  });
+});
